Search with "AND logic" rather than "OR logic"
Adds a variable to the search function, filtered_rows, which is the dataframe which goes through the actual filtering if/else logic and comes back to the top of the "for" loop to be re-filtered by further search terms. It is these rows which are then whittled down to just EntryNum and then rejoined back to the original filtered_data dataframe.